Move snippets for properties article
