+++ docs/csharp/programming-guide/classes-and-structs/snippets/how-to-declare-and-use-read-write-properties/Properties.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Properties</RootNamespace>
+  </PropertyGroup>
+
+</Project>
