+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideObjects/CS/Objects.cs
@@ -802,109 +802,6 @@ class TestPerson
     //</Snippet32>
 
     //-----------------------------------------------------------------------------
-    namespace WrapPerson2
-    {
-        //<Snippet33>
-        class Person
-        {
-            private string _name = "N/A";
-            private int _age = 0;
-
-            // Declare a Name property of type string:
-            //<Snippet34>
-            public string Name
-            {
-                get
-                {
-                    return _name;
-                }
-                set
-                {
-                    //<Snippet36>
-                    _name = value;
-                    //</Snippet36>
-                }
-            }
-            //</Snippet34>
-
-            // Declare an Age property of type int:
-            public int Age
-            {
-                get
-                {
-                    return _age;
-                }
-
-                set
-                {
-                    _age = value;
-                }
-            }
-
-            //<Snippet38>
-            public override string ToString()
-            {
-                return "Name = " + Name + ", Age = " + Age;
-            }
-            //</Snippet38>
-        }
-
-        class TestPerson
-        {
-            static void Main()
-            {
-                // Create a new Person object:
-                Person person = new Person();
-
-                // Print out the name and the age associated with the person:
-                Console.WriteLine("Person details - {0}", person);
-
-                // Set some values on the person object:
-                //<Snippet35>
-                person.Name = "Joe";
-                person.Age = 99;
-                //</Snippet35>
-                Console.WriteLine("Person details - {0}", person);
-
-                // Increment the Age property:
-                //<Snippet37>
-                person.Age += 1;
-                //</Snippet37>
-                Console.WriteLine("Person details - {0}", person);
-
-                // Keep the console window open in debug mode.
-                Console.WriteLine("Press any key to exit.");
-                Console.ReadKey();
-            }
-        }
-        /* Output:
-            Person details - Name = N/A, Age = 0
-            Person details - Name = Joe, Age = 99
-            Person details - Name = Joe, Age = 100
-        */
-        //</Snippet33>
-    }
-
-    //-----------------------------------------------------------------------------
-    namespace WrapPerson4
-    {
-        class Person
-        {
-            private string _name = "N/A";
-
-            //<Snippet87>
-            public string Name
-            {
-                get
-                {
-                    return _name;
-                }
-            }
-            //</Snippet87>
-        }
-    }
-
-    //-----------------------------------------------------------------------------
     //<Snippet79>
     public class Customer
     {
